requirements.txt
streamlit
langchain
langchain_groq
python-dotenv
groq
